Added more efficiency and bugfixes for difference reporting.

+++ src/facts/diff/BasicDelete.java
@@ -12,7 +12,7 @@ public class BasicDelete extends TreeEditOperation
 	      super.opName = "DELETE";
     }
 
-    public double getCost(int aNodeID, int bNodeID, TreeNode aTree, TreeNode bTree) 
+    public Integer getCost(int aNodeID, int bNodeID, TreeNode aTree, TreeNode bTree) 
     {
 	      return 1;
     }
